+++ tests/testthat/test-UploadPlpDbResults_upadted.R
@@ -60,8 +60,13 @@ test_that("results uploaded to database", {
   if(!dir.exists(file.path(resultsLoc,'Validation','test', 'Analysis_1'))){
     dir.create(file.path(resultsLoc,'Validation','test', 'Analysis_1'), recursive = T)
   }
-  plpResult$model$valdiationDetails <- list(cohortId = 1, outcomeId = 3)
-  saveRDS(plpResult, file.path(resultsLoc,'Validation','test', 'Analysis_1', 'validationResult.rds'))
+  plpResult$model$validationDetails <- list(
+    cohortId = 1, 
+    outcomeId = 3,
+    populationSettings = plpResult$model$settings$populationSettings, 
+    plpDataSettings = plpResult$model$settings$plpDataSettings
+    )
+  savePlpResult(plpResult, file.path(resultsLoc,'Validation','test', 'Analysis_1', 'validationResult'))
 
   # add results:
   populatePlpResultTables(conn = conn,